Don't add association filters by default
First of all, let me say that I haven't found any issues related to this
(I might haved missed something), so if there's already a solution or an
alternative being discussed, apologies.

My current use case is this:
I have a model

```
Edition.belongs_to Single
```

There are about 34k editions on my database, and 19k singles.

Since ex_admin creates a filters for all associations by default, on
`/admin/editions`, I get a dropdown with 19k items on it.
This is not only a useless filter, but it also requires a query that
takes more than 1s to run.

After digging a lot through the code and the docs, I couldn't find a way
to disable this field (PS: the docs for `filter` could perhaps be
improved to explain that these association filters are not
configurable)

I ended up with a solution that works for me in the meantime, and might
as well be used to start discussion here: remove the association filters
altogether.

My reasoning is not only that they are dangerous for performance in
cases such as mine (I imagine having a `belongs_to` with thousands of
records is not such a rare thing to have).
It also comes from the fact that it's pretty simple to add filters for
the desired associations:

```
filter [:name, :single]
```

I'm not sure if this goes against what the author intends for the
default functionality to be, and if so, I'll gladly discuss other
solutions.