docs: added new use case for Register.after_filter
This commit updates the docs of Register.after_filter.
It explains how to use this filter to add new users
to a web app using ExAdmin.

While this might a be a little too specific, it is a common use case.
The documentation on the filter (while providing a helpful code
example) could use some prose describing why you would want to use it
in practice.